Move python into docker container
Communicating back and forth between the docker container and the host
was proven difficult and awkward. Especially when I am generating the
puzzles, I need to be able to run a bunch of file operations natively.
Running exec_commands was also relatively slow.

I've stuck Python in the container now, and it can run the generators.
And since I'm doing that we might as well run the GUI in the container as
well. Apparently that is possible with Docker with a few tweaks.

All this does increase the docker image size substantially unfortunately.
I may try to shrink it some later.