UserCodeError stores traceback as string now
Storing the original exception is risky, since the user may have some
form of custom exception or something from a library that may not pickle,
or the exception class may not exist outside of the container. It is safer
to send the traceback info as a string. Its all we actually want anyways.