Pretty traceback messages
Since we don't want the puzzles to be saved to disk in the container,
we've been executing the puzzles as a string. But `exec`ed strings don't
get as much traceback info when an exception is thrown in the code.

Now the tutorial magically puts that info back in the traceback if user
code such as puzzles throws an error. This allows us to print the exception
and show lines and the file on the host that the error occurred in.