Remove undo functionality
Since reloading a Docker commit required restarting the container, undo
would make you lose any in-memory state such as environment variables,
running processes, etc, which could cause lots of issues.

Now, we only keep the functionality for restarting to the beginning of the
tutorial. We use a docker commit so we can restore to the same tutorial
state on a restart in case of randomized puzzles.