Get coverage working
We can now view code coverage. We can combine coverage reports from the
docker_tests and the main tests into one report.

Coverage reports are somewhat incomplete as we don't have a good way to
track the code that gets run in the container during integration tests,
since the container is launched and closed during the test.