Restructure api package and RandomHelper to reduce circular dependencies
The fact that the api package was using a global _tutorial variable that
referenced TutorialDocker was causing issues. In order to reduce the
dependencies, RandomHelper is now in the api package. And instead of having
a global _tutorial variable, the api has _home and _rand variables which
will be set by the TutorialDocker.

I wish there was a better way to pas the Tutorial's RandomHelper and home
to the api than having to set these globals.