Add PuzzleData class
Separate Puzzle into two classes. Puzzle is the pure data class which
will be available to the user to create puzzles. We will then convert it
into a PuzzleData object which contains extra metadata and handles
pickling and stuff. This way we don't have our fields showing up on the
class that the user is going to be using.