Remove umask
Setting the umask caused files to be created with different permissions
in the templates than when made by the student.

Pathlib's touch apparently uses the umask even if you set the `mode` option,
which was why we set `umask` to 0o000 in the first place. To fix that just
use chmod() in the create method.