Move Puzzle into shared
To make the project run on Windows, we need to make sure the Docker side
modules do not get imported host side. Since we imported api.Puzzle in shared,
and imported shared in host, we ended up importing the api modules on host side.
Move Puzzle into shared fixes the issue so we only load the modules we actually
need on host.