Don't pull image if exists locally
Even though pull shouldn't download anything if the image already exists
locally and hasn't changed, the pull takes quite a while. Especially on
Docker for Windows.

So, only pull if the image doesn't exist locally. This does mean that if
you update the Docker container Shell Adventure won't update to the newer
container. We'll just have to rely on docker container tags to switch to
the newer container.