Get tests to pass on Windows
Fix various issues with running on Windows.

Also, it appears that the Docker API hangs if you pass it a malformed
request on Docker for Windows. This caused one of my tests of incorrect
container_options to hang. I've skipped that test on windows, and add
docs that container_options is "use at your own risk"